Several changes and updates
Updated - Node.h, Primitive.h, SceneGraph.h, SceneValues.h,
TransformValues.h, XMLScene.h, Node.cpp, SceneGraph.cpp,
TransformValues.cpp, XMLScene.cpp

-Corrected TransformValues and implemented its apply() method;
-Small corrections to Primitive.h and SceneValues.h
-Added printNodesID() method to SceneGraph

-Updated Node.draw() and XMLScene.drawScene(Node*) - still has bugs and
not fully implemented

+++ 3D Graphics Engine/include/TransformValues.h
@@ -2,6 +2,7 @@
 #define TRANSFORMVALUES_H
 
 #include "SceneValues.h"
+#include "CGFobject.h"
 
 #include <map>
 #include <string>
@@ -14,9 +15,9 @@ class TransformValues : public SceneValues
 {
 private:
 
-	map<string, double> scale;
-	map<string, double> translate;
-	map<string, double> rotate;
+	map<string, float> scale;
+	map<string, float> translate;
+	map<string, float> rotate;
 
 public:
 	TransformValues();
